Update README.md
The dataset used for this analysis includes key indicators such as depression, anxiety, and panic attacks, along with demographic and academic details. The objective of this study is to identify trends, uncover risk factors, and provide insights into the prevalence of mental health issues among students. It aims to contribute to better mental health awareness and support strategies within academic environments.